// packages/monaco-lsp-client/src/types.ts
export enum TokenFormat {
  Relative = "relative",
}
// ...
export type DocumentSelector = DocumentFilter[];
// ...
export interface WorkDoneProgressOptions {
  workDoneProgress?: boolean;
}

/**
 * General text document registration options.
 */
export interface TextDocumentRegistrationOptions {
  /**
   * A document selector to identify the scope of the registration. If set to null
   * the document selector provided on the client side will be used.
   */
  documentSelector: DocumentSelector | null;
}
// ...
export interface SemanticTokensRegistrationOptions
  extends TextDocumentRegistrationOptions, SemanticTokensOptions, StaticRegistrationOptions {}
// ...
export interface StaticRegistrationOptions {
  /**
   * The id used to register the request. The id can be used to deregister
   * the request again. See also Registration#id.
   */
  id?: string;
}
// ...
export interface SemanticTokensOptions extends WorkDoneProgressOptions {
  /**
   * The legend used by the server
   */
  legend: SemanticTokensLegend;
  /**
   * Server supports providing semantic tokens for a specific range
   * of a document.
   */
  range?: boolean | Record<string, never>;
  /**
   * Server supports providing semantic tokens for a full document.
   */
  full?: boolean | { delta?: boolean };
}
// ...
export interface SemanticTokensLegend {
  /**
   * The token types a server uses.
   */
  tokenTypes: string[];
  /**
   * The token modifiers a server uses.
   */
  tokenModifiers: string[];
}